tests/wpt/web-platform-tests/workers/constructors/SharedWorker/unexpected-global-properties.html

<!doctype html>
<title>unexpected members/interface objects/constructors</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
(async_test()).step(function() {
  var worker = new SharedWorker('unexpected-global-properties.js');
  worker.port.addEventListener('message', this.step_func(function(e) {
    assert_equals(e.data, '');
    this.done();
  }), false);
  worker.port.start();
});
</script>
